Added infos about ram quantity and debug log

+++ srcs/lib/intlib.h
@@ -6,14 +6,14 @@
 /*   By: lubenard <lubenard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 17:38:00 by lubenard          #+#    #+#             */
-/*   Updated: 2022/03/08 17:57:45 by lubenard         ###   ########.fr       */
+/*   Updated: 2022/08/18 23:09:36 by luca             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef INTLIB_H
 # define INTLIB_H
 
-int intlen(int n);
+int intlen(long n);
 int uintlen(unsigned int n);
 
 #endif
